Add files via upload
This file contains the EDA and Sentiment analysis for Airbnb - Seattle dataset. I have worked upon the part of sentiment analysis where I am trying to verify if the ratings and the neighborhood align with each other. If the neighborhood is giving a positive vibe then it is described using words like good, best, great, enjoy. I have used NLTK  library of python. I have also used word cloud to identify the keywords from the neighborhood overview so that we see what words describe the vibe of Seattle neighborhood.